style: fix all clippy warnings reported by CI lint job
Address 26 clippy lints (all -D warnings level) to make the lint job
green without weakening the CI's strictness. No behavioural changes.

Mechanical idiomatic rewrites:
- &PathBuf -> &Path in cgi_handler, error_page_handler signatures
- push_str("X") -> push('X') in directory_listing_handler
- or_insert_with(Vec::new) -> or_default() in validator, headers
- format!("static") -> .to_string() in redirection_handler, server_manager
- .clone() on Copy ConnectionState -> *deref in server_manager
- code >= 100 && code <= 599 -> (100..=599).contains(&code) in status
- (EV_ADD | EV_ENABLE) as u16 -> drop cast (already u16) in poller
- trim().split_whitespace() -> split_whitespace() in cgi_io
- is_some() + unwrap() -> if let Some(...) in server_manager
- manual &path[2..] after starts_with -> strip_prefix in router
- CLEANUP_COUNTER % 100 == 0 -> .is_multiple_of(100) in server_manager

Refactors required by stricter lints:
- impl Default for RouteConfig -> #[derive(Default)] in models
- impl Default for Version -> #[derive(Default)] + #[default] in version
- RouteConfig::default() + field reassign -> struct literal in router tests
- contains_key + insert -> Entry::Vacant/Occupied in server_manager
- match { block } -> let res = { block }; match res in server_manager (2x)
- pub fn Headers::to_string -> impl Display (kept .to_string() via blanket impl)

module_inception on src/core/event/event.rs is silenced with a local
#[allow] on the parent mod since renaming would break the public path.

+++ src/core/event/poller.rs
@@ -48,7 +48,7 @@ impl Poller {
             let kev = KeventStruct {
                 ident: fd as usize,
                 filter: filter as i16,
-                flags: (EV_ADD | EV_ENABLE) as u16,
+                flags: EV_ADD | EV_ENABLE,
                 fflags: 0,
                 data: 0,
                 udata: user_data as *mut c_void,
@@ -77,7 +77,7 @@ impl Poller {
             let kev = KeventStruct {
                 ident: fd as usize,
                 filter: filter as i16,
-                flags: EV_DELETE as u16,
+                flags: EV_DELETE,
                 fflags: 0,
                 data: 0,
                 udata: std::ptr::null_mut(),
